some updates...

+++ source/magicprocessing/packetjumper.d
@@ -26,8 +26,8 @@ class PacketJumper : Processing {
 
 		auto newMethodContent = appender!string();
 
-		int local2Count = count(sendHelloMethod.content, "getlocal2");
-		int local2Found = 0;
+		ulong local2Count = count(sendHelloMethod.content, "getlocal2");
+		ulong local2Found = 0;
 
 		foreach(string line; util.string.readLines(sendHelloMethod.content)) {
 			if(stat != patchStat.success) {
